fit_transform issue

+++ lightning/primal_cd.py
@@ -61,7 +61,7 @@ class PrimalLinearSVC(BaseSVC, BaseLinearClassifier, ClassifierMixin):
         X = np.asfortranarray(X, dtype=np.float64)
 
         self.label_binarizer_ = LabelBinarizer(neg_label=-1, pos_label=1)
-        Y = self.label_binarizer_.fit_transform(y)
+        Y = self.label_binarizer_.fit_transform(y).astype(np.float64)
         self.classes_ = self.label_binarizer_.classes_
         n_vectors = Y.shape[1]
 
@@ -137,7 +137,7 @@ class PrimalSVC(BaseSVC, BaseKernelClassifier, ClassifierMixin):
         X = np.ascontiguousarray(X, dtype=np.float64)
 
         self.label_binarizer_ = LabelBinarizer(neg_label=-1, pos_label=1)
-        Y = self.label_binarizer_.fit_transform(y)
+        Y = self.label_binarizer_.fit_transform(y).astype(np.float64)
         self.classes_ = self.label_binarizer_.classes_.astype(np.int32)
         n_vectors = Y.shape[1]
 
@@ -251,7 +251,7 @@ class PrimalL2SVC(BaseSVC, BaseKernelClassifier, ClassifierMixin):
         X = np.ascontiguousarray(X, dtype=np.float64)
 
         self.label_binarizer_ = LabelBinarizer(neg_label=-1, pos_label=1)
-        Y = self.label_binarizer_.fit_transform(y)
+        Y = self.label_binarizer_.fit_transform(y).astype(np.float64)
         self.classes_ = self.label_binarizer_.classes_.astype(np.int32)
         n_vectors = Y.shape[1]
 
@@ -279,7 +279,7 @@ class PrimalL2SVC(BaseSVC, BaseKernelClassifier, ClassifierMixin):
 
 def C_lower_bound(X, y, kernel=None, search_size=None, random_state=None,
                   **kernel_params):
-    Y = LabelBinarizer(neg_label=-1, pos_label=1).fit_transform(y)
+    Y = LabelBinarizer(neg_label=-1, pos_label=1).fit_transform(y).astype(np.float64)
 
     if kernel is None:
         den = np.max(np.abs(np.dot(Y.T, X)))
